ml-backend/requirements.txt
Flask==2.3.1
pillow==10.2.0
requests==2.31.0
numpy
keras-nightly
flask_cors